src/models/train.py
# Model Training
# This module contains model training functionality

import pandas as pd
import numpy as np
from typing import Tuple, Dict, Any
import joblib
from datetime import datetime


def train_model(X: pd.DataFrame, y: pd.Series) -> Tuple[Any, Dict[str, Any]]:
    """
    Train the fraud detection model
    
    Args:
        X: Feature matrix
        y: Target vector
        
    Returns:
        Tuple of (trained_model, metrics)
    """
    # TODO: Implement actual model training
    # Placeholder for model training logic
    
    from sklearn.ensemble import RandomForestClassifier
    
    model = RandomForestClassifier(
        n_estimators=100,
        max_depth=10,
        random_state=42,
        class_weight="balanced"
    )
    
    model.fit(X, y)
    
    # Calculate metrics
    train_score = model.score(X, y)
    
    metrics = {
        "train_accuracy": train_score,
        "timestamp": datetime.utcnow().isoformat(),
        "n_features": X.shape[1],
        "n_samples": X.shape[0]
    }
    
    return model, metrics


def evaluate_model(model: Any, X: pd.DataFrame, y: pd.Series) -> Dict[str, float]:
    """
    Evaluate model performance
    
    Args:
        model: Trained model
        X: Feature matrix
        y: Target vector
        
    Returns:
        Dictionary of evaluation metrics
    """
    from sklearn.metrics import accuracy_score, precision_score, recall_score, f1_score
    
    y_pred = model.predict(X)
    
    metrics = {
        "accuracy": accuracy_score(y, y_pred),
        "precision": precision_score(y, y_pred, zero_division=0),
        "recall": recall_score(y, y_pred, zero_division=0),
        "f1_score": f1_score(y, y_pred, zero_division=0)
    }
    
    return metrics


def save_model(model: Any, filepath: str) -> None:
    """
    Save model to file
    
    Args:
        model: Trained model
        filepath: Path to save model
    """
    joblib.dump(model, filepath)


def load_model(filepath: str) -> Any:
    """
    Load model from file
    
    Args:
        filepath: Path to model file
        
    Returns:
        Loaded model
    """
    return joblib.load(filepath)
