requirements.txt
# Core Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Deep Learning
tensorflow>=2.14.0

# Explainability
shap>=0.43.0

# API Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security & Authentication
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.2.0

# Configuration & Environment
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Caching & Storage
redis>=5.0.0

# Dashboard
streamlit>=1.33.0

# Visualization
matplotlib>=3.9.0
seaborn>=0.13.0

# Data Generation
faker>=30.0.0

# HTTP Client
requests>=2.31.0
httpx>=0.26.0

# Utilities
joblib>=1.3.0
python-dateutil>=2.9.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development Tools
jupyter>=1.0.0
ipython>=8.27.0
notebook>=7.2.0

# Code Quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
